requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
axios==0.4.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
dj-static==0.0.6
Django==4.2.4
django-cors-headers==3.4.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
gunicorn==21.2.0
h5py==3.9.0
idna==3.4
keras==2.13.1
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
protobuf==4.24.2
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rsa==4.9
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.12.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflowjs==3.18.0
termcolor==2.3.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
whitenoise==6.5.0
wrapt==1.15.0
